+++ src/main/java/com/github/nkoutroumanis/histogram/Space2D.java
@@ -2,35 +2,35 @@ package com.github.nkoutroumanis.histogram;
 
 public final class Space2D {
 
-    private final float maxx;
-    private final float maxy;
-    private final float minx;
-    private final float miny;
+    private final double maxx;
+    private final double maxy;
+    private final double minx;
+    private final double miny;
 
-    private Space2D(float minx, float miny, float maxx, float maxy) {
+    private Space2D(double minx, double miny, double maxx, double maxy) {
         this.minx = minx;
         this.miny = miny;
         this.maxx = maxx;
         this.maxy = maxy;
     }
 
-    public static Space2D newSpace2D(float minx, float miny, float maxx, float maxy) {
+    public static Space2D newSpace2D(double minx, double miny, double maxx, double maxy) {
         return new Space2D(minx, miny, maxx, maxy);
     }
 
-    public float getMaxx() {
+    public double getMaxx() {
         return maxx;
     }
 
-    public float getMaxy() {
+    public double getMaxy() {
         return maxy;
     }
 
-    public float getMinx() {
+    public double getMinx() {
         return minx;
     }
 
-    public float getMiny() {
+    public double getMiny() {
         return miny;
     }
 }
